add he possibilty in the count routine to not only increment the nucleotide per position by 1, but by the given number times; in utils, save id as long instead of int, they can get quite large with large sequence length L

+++ src/utils.hpp
@@ -38,7 +38,7 @@ std::string::iterator find_tab_in_string(std::string::iterator it)
     return it;
 }
 
-int choose(const int n, const int k);
+unsigned long choose(const int n, const int k);
 
 }
